requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
celery[redis]==5.4.0
redis==5.2.1
python-multipart==0.0.19
PyMuPDF==1.25.3
python-docx==1.1.2
Pillow==11.1.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
aiofiles==24.1.0
